refactor: exception.error 패키지로 이동
