+++ src/main/java/com/example/demo/exception/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.example.demo.exception;
+package com.example.demo.exception.error;
 
 import org.springframework.http.HttpStatus;
