+++ src/main/java/com/example/demo/exception/error/MemberErrorCode.java
@@ -1,15 +1,16 @@
-package com.example.demo.exception;
+package com.example.demo.exception.error;
 
 import org.springframework.http.HttpStatus;
 
-public enum BoardErrorCode implements ErrorCode{
+public enum MemberErrorCode implements ErrorCode{
 
-    ARTICLE_EXISTENCE(HttpStatus.BAD_REQUEST, "작성된 게시물이 존재합니다.");
+    EMAIL_CONFLICT(HttpStatus.CONFLICT, "중복된 이메일입니다."),
+    ARTICLE_EXISTENCE(HttpStatus.BAD_REQUEST, "작성한 게시물이 존재합니다.");
 
     private final HttpStatus httpStatus;
     private final String message;
 
-    BoardErrorCode(HttpStatus httpStatus, String message) {
+    MemberErrorCode(HttpStatus httpStatus, String message) {
         this.httpStatus = httpStatus;
         this.message = message;
     }
